Updating the element observer.
Must figure out a way to prevent from creating a new intersection observer for every element passed. Is this inefficient?
